Switch JUnit 4 example to new agent
