+++ pom.xml
@@ -25,6 +25,9 @@
     </modules>
 
     <properties>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<maven.compiler.target>1.8</maven.compiler.target>
+		<maven.compiler.source>1.8</maven.compiler.source>
         <slf4jVersion>1.7.25</slf4jVersion>
         <log4j-core.version>2.10.0</log4j-core.version>
     </properties>
@@ -55,22 +58,6 @@
             <version>${slf4jVersion}</version>
         </dependency>
 
-
     </dependencies>
-
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.0</version>
-                <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
 
 </project>
