--- v4/src/main/java/com/epam/rp/tests/SimpleTest1.java
@@ -1,11 +1,0 @@
-package com.epam.rp.tests;
-
-import org.junit.Test;
-
-public class SimpleTest1 {
-
-	@Test
-	public void testPrintSomething() {
-		System.out.println("test123");
-	}
-}
