--- v4/src/main/java/com/epam/rp/tests/SimpleTest2.java
@@ -1,10 +1,0 @@
-package com.epam.rp.tests;
-
-import org.junit.Test;
-
-public class SimpleTest2 {
-	@Test
-	public void testPrintSomething() {
-		System.out.println("test123-2");
-	}
-}
