--- v4/src/main/java/com/epam/rp/tests/SimpleTestSuite1.java
@@ -1,10 +1,0 @@
-package com.epam.rp.tests;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-
-@RunWith(MySuite.class)
-@Suite.SuiteClasses({ SimpleTest1.class, SimpleTest2.class })
-public class SimpleTestSuite1 {
-
-}
