+++ v4/src/test/java/com/epam/rp/tests/TestIssue4.java
@@ -1,0 +1,17 @@
+package com.epam.rp.tests;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+public class TestIssue4 {
+
+	@Test
+	public void test1() {
+		Assert.assertTrue("First result", true);
+	}
+
+	@Test
+	public void test2() {
+		Assert.assertTrue("Second result", true);
+	}
+}
