+++ v4/src/test/java/com/epam/rp/tests/TestSimple1.java
@@ -1,0 +1,12 @@
+package com.epam.rp.tests;
+
+import org.junit.Test;
+
+public class TestSimple1 {
+
+	@Test
+	public void testPrintSomething() {
+		System.out.println("test123");
+	}
+	
+}
