+++ v4/src/test/java/com/epam/rp/tests/TestSimple2.java
@@ -1,0 +1,12 @@
+package com.epam.rp.tests;
+
+import org.junit.Test;
+
+public class TestSimple2 {
+	
+	@Test
+	public void testPrintSomething() {
+		System.out.println("test123-2");
+	}
+	
+}
