+++ v4/src/test/java/com/epam/rp/tests/TestSimpleSuite.java
@@ -1,0 +1,11 @@
+package com.epam.rp.tests;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+import org.junit.runners.Suite.SuiteClasses;
+
+@RunWith(Suite.class)
+@SuiteClasses({ TestSimple1.class, TestSimple2.class })
+public class TestSimpleSuite {
+
+}
